02-ddp-pytorch-ray/version2-with-ray-data/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ray[train,data]>=2.9.0
tqdm
pandas
pillow
